refactor: move CourseCompetencyProgressDTO to atlas module
# Conflicts:
#	src/main/java/de/tum/cit/aet/artemis/atlas/service/competency/CompetencyProgressService.java
#	src/main/java/de/tum/cit/aet/artemis/atlas/service/competency/CourseCompetencyService.java
#	src/main/java/de/tum/cit/aet/artemis/atlas/web/CourseCompetencyResource.java
#	src/test/java/de/tum/cit/aet/artemis/atlas/competency/CourseCompetencyIntegrationTest.java
