+++ src/main/java/de/tum/cit/aet/artemis/atlas/dto/CourseCompetencyProgressDTO.java
@@ -1,0 +1,10 @@
+package de.tum.cit.aet.artemis.atlas.dto;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+/**
+ * DTO representing the competency progress of the whole course.
+ */
+@JsonInclude(JsonInclude.Include.NON_EMPTY)
+public record CourseCompetencyProgressDTO(long competencyId, long numberOfStudents, long numberOfMasteredStudents, double averageStudentScore) {
+}
